// eMAR from Winston Laptop/Presentation/A_Production/emar/src/app/interfaces/user-remembered-list-order.ts
export interface UserRememberedListOrder {
  userId: number;
  siteId: number;
  medicationId: number;
  dose?: number;
  medicationUnitId?: number;
  medicationRouteId?: number;
  priority?: string;
  frequencyId?: number;
  orderNotes?: string;
  duration?: number;
  durationUnitId?: number;
  ndc?: string;
}